* release:
  Bump timeout on TimeoutReader test because it's close
  Bump timeout on TimeoutReader test because it's close
  Clean up queue handling in GlobalOp to produce closer to expected results for Discard on queue full
  Tweak SST Data transport so that if a transport is unavailable, we don't use it even if preferred.  Set SST Control transport to default to sockets.  Establish nicer aliases for Control transports.  Document control transport specification.
  add cast
  add cast
  add cast
  add cast
  Remove usleep for MSVC
  Remove usleep for MSVC
  Try to fix VC++ compilation.
  Remove KillReader tests until stabilized.  Try to fix VC++ compilation.
  Restore SST no-mpi testing by letting staging-common activate on a no-mpi build
  clang-format
  Sort variable and attributes names to return the same list as the C++/Python API in adios2_inquire_all_variables() and adios2_inquire_all_attributes()
  Support option 'accuracy' in MGARD, as a synonym for tolerance

+++ testing/adios2/engine/staging-common/TestCommonReadAttrs.cpp
@@ -116,7 +116,7 @@ TEST_F(CommonReadTest, ADIOS2CommonRead1D8)
         const size_t currentStep = engine.CurrentStep();
         EXPECT_EQ(currentStep, static_cast<size_t>(t));
 
-        int writerSize;
+        size_t writerSize;
 
         auto attr_s1 = io.InquireAttribute<std::string>(s1_Single);
         auto attr_s1a = io.InquireAttribute<std::string>(s1_Array);
@@ -278,13 +278,14 @@ TEST_F(CommonReadTest, ADIOS2CommonRead1D8)
         ASSERT_EQ(var_time.ShapeID(), adios2::ShapeID::GlobalArray);
         ASSERT_EQ(var_time.Shape()[0], writerSize);
 
-        long unsigned int myStart = (writerSize * Nx / mpiSize) * mpiRank;
+        long unsigned int myStart =
+            (long unsigned int)(writerSize * Nx / mpiSize) * mpiRank;
         long unsigned int myLength =
-            ((writerSize * Nx + mpiSize - 1) / mpiSize);
+            (long unsigned int)((writerSize * Nx + mpiSize - 1) / mpiSize);
 
         if (myStart + myLength > writerSize * Nx)
         {
-            myLength = writerSize * Nx - myStart;
+            myLength = (long unsigned int)writerSize * Nx - myStart;
         }
         const adios2::Dims start{myStart};
         const adios2::Dims count{myLength};
